docs: add debian/ubuntu links to README.md
The Debian CI site is apparently used for testing packages before the
official Debian packages are built.

Currently it is only listed in places like #6702, so add it to the main
list to make it easier to find.

Also add links for the Debian Package Tracker and Ubuntu PPA since they
are related.

Relates to #6702 #6842.