bugfix: do sandbox check before checkcfg()
The check for an existing sandbox (and running `run_no_sandbox()` if
applicable) must be done before calling `checkcfg()`, since if
`private-etc` is already in effect, running firejail again will abort at
`checkcfg()`, as /etc/firejail/firejail.config will not be accessible
(see #6966).

This is a follow-up to #6969.

Relates to #2877 #6878 #6951 #6966.

Kind of relates to #6592.