docs: github: specify the supported version directly
Bug reports were opened by multiple users with the item checked even
when using version 0.9.74 or earlier.

So be explicit and write "0.9.76 or later" in the item, which should
help clarify that earlier versions are not supported.

This amends commit be065d130 ("docs: clarify that only latest and dev
versions are supported (#6965)", 2025-11-15).