firecfg: restore removed messages but make them debug
They are useful for debugging during development, especially the
`removed` message.

The `GLOB_NOMATCH` message should never happen with the current code,
which could be used to check for regressions.  It is also consistent
with the other checks for `GLOB_NOMATCH` and should be useful if
user-supplied globs are later supported.

This amends commit 4124e76b4 ("firecfg: Ubuntu snap integration",
2025-12-27).